migrate tailwind-gridflex to shared endpoint
