+++ pages/api/studentPortal/lessons/web/react/index.ts
@@ -2,6 +2,7 @@ import { getDataForComponentsLesson } from "./components";
 import { getGithubLessonData } from "./github";
 import { getPropsLessonData } from "./props";
 import { getTailwindColourStylingLessonData } from "./tailwindcss-colourstyling";
+import { getTailwindGridFlexLessonData } from "./tailwindcss-gridflex";
 
 export function getLessonForReactCourse(lessonId: string) {
   if (lessonId === "components") {
@@ -10,8 +11,10 @@ export function getLessonForReactCourse(lessonId: string) {
     return getGithubLessonData();
   } else if (lessonId === "props") {
     return getPropsLessonData();
-  }else if (lessonId === "tailwindcss-colourstyling") {
+  } else if (lessonId === "tailwindcss-colourstyling") {
     return getTailwindColourStylingLessonData();
+  } else if (lessonId === "tailwindcss-gridflex") {
+    return getTailwindGridFlexLessonData();
   }
 
   throw new Error(`Could not find lessons`);
